proximity sensor

+++ src/org/usfirst/frc/team5461/sensors/VL6180xIdentification.java
@@ -2,11 +2,11 @@ package org.usfirst.frc.team5461.sensors;
 
 public class VL6180xIdentification
 {
-  public byte idModel;
-  public byte idModelRevMajor;
-  public byte idModelRevMinor;
-  public byte idModuleRevMajor;
-  public byte idModuleRevMinor;
+  public int idModel;
+  public int idModelRevMajor;
+  public int idModelRevMinor;
+  public int idModuleRevMajor;
+  public int idModuleRevMinor;
   public short idDate;
   public short idTime;
 }
